Split apart the README file and update Contributing file from the wiki
This includes migrating the following pages from the wiki to
the source:

- https://wiki.eclipse.org/CDT/contributing
- https://wiki.eclipse.org/Getting_started_with_CDT_development
- https://wiki.eclipse.org/CDT/policy
- https://wiki.eclipse.org/CDT/git

Note that much of the content on those pages was redundant or
obsolete, therefore the files are not copied over as is
from the wiki.

In addition the content from https://www.eclipse.org/cdt/ has
migrated over as appropriate too. Much of https://www.eclipse.org/cdt/
was also redundant or obsolete too.

Change-Id: I2b38a4fbb3351c8eb8bbf8c910eb3a3118eda142